moved github logic to service
