+++ CodeConnect.sln.DotSettings.user
@@ -55,5 +55,6 @@
     &lt;/TestAncestor&gt;
   &lt;/Or&gt;
 &lt;/SessionState&gt;</s:String>
+	<s:Boolean x:Key="/Default/UnloadedProject/UnloadedProjects/=3d8bf299_002D5e11_002D4a03_002Db03b_002D00f065db7ff2_0023ApiServiceTests/@EntryIndexedValue">True</s:Boolean>
 	
 	</wpf:ResourceDictionary>
