+++ DomainLayer/Entities/APIClasses/GithubSettings.cs
@@ -1,0 +1,7 @@
+namespace DomainLayer.Entities.APIClasses;
+
+public class GithubSettings
+{
+    public string ClientId { get; set; } = "";
+    public string Secret { get; set; } = "";
+}
